requirements.txt
fastapi
uvicorn
pymongo
langchain
langchain-community
python-dotenv
pydantic
bson